service com crud básico feito, junto com algumas pequenas alterações de sintax

+++ restaurante/src/main/java/com/anagabriella/restaurante/repository/ItemPedidoRepository.java
@@ -2,5 +2,5 @@ package com.anagabriella.restaurante.repository;
 import com.anagabriella.restaurante.entity.ItemPedido;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ItemPedidoRepository extends JpaRepository<ItemPedido, Long> {
+public interface ItemPedidoRepository extends JpaRepository<ItemPedido, Integer> {
 }
