+++ restaurante/src/main/java/com/anagabriella/restaurante/repository/PedidoRepository.java
@@ -2,5 +2,5 @@ package com.anagabriella.restaurante.repository;
 import com.anagabriella.restaurante.entity.Pedido;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PedidoRepository extends JpaRepository<Pedido, Long> {
+public interface PedidoRepository extends JpaRepository<Pedido, Integer> {
 }
